remove common runtime and refactor pallet configs
