+++ Cargo.toml
@@ -4,7 +4,6 @@ panic = 'unwind'
 [workspace]
 members = [
   'node',
-  'runtime/common',
   'runtime/basilisk',
   'pallets/marketplace',
   'pallets/xyk',
