+++ node/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "basilisk"
-version = "10.0.0"
+version = "10.0.1"
 description = "Basilisk node"
 authors = ["GalacticCouncil"]
 edition = "2021"
@@ -29,7 +29,6 @@ clap = { version = "4.1.8", features = [ "derive" ] }
 futures = "0.3.21"
 
 # local dependencies
-common-runtime = { path = "../runtime/common" }
 basilisk-runtime = { path = "../runtime/basilisk" }
 primitives = { path = "../primitives" }
