+++ primitives/src/lib.rs
@@ -26,13 +26,47 @@ use serde::{Deserialize, Serialize};
 
 use scale_info::TypeInfo;
 
-use frame_support::{parameter_types, sp_runtime::FixedU128};
+use frame_support::{
+	parameter_types,
+	sp_runtime::FixedU128,
+	sp_runtime::{
+		generic,
+		traits::{BlakeTwo256, IdentifyAccount, Verify},
+		MultiSignature, Perbill,
+	},
+};
 
 pub mod asset;
 pub mod constants;
 
 /// An index to a block.
 pub type BlockNumber = u32;
+
+/// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
+pub type Signature = MultiSignature;
+
+/// Some way of identifying an account on the chain. We intentionally make it equivalent
+/// to the public key of our transaction signing scheme.
+pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;
+
+/// The type for looking up accounts. We don't expect more than 4 billion of them, but you
+/// never know...
+pub type AccountIndex = u32;
+
+/// Index of a transaction in the chain.
+pub type Index = u32;
+
+/// A hash of some data used by the chain.
+pub type Hash = sp_core::H256;
+
+/// Opaque, encoded, unchecked extrinsic.
+pub use sp_runtime::OpaqueExtrinsic as UncheckedExtrinsic;
+
+/// Header type.
+pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
+
+/// Block type.
+pub type Block = generic::Block<Header, UncheckedExtrinsic>;
 
 /// Type used for expressing timestamp.
 pub type Moment = u64;
@@ -54,6 +88,13 @@ pub type CollectionId = u128;
 
 /// NFT Item ID
 pub type ItemId = u128;
+
+/// We assume that an on-initialize consumes 2.5% of the weight on average, hence a single extrinsic
+/// will not be allowed to consume more than `AvailableBlockRatio - 2.5%`.
+pub const AVERAGE_ON_INITIALIZE_RATIO: Perbill = Perbill::from_perthousand(25);
+/// We allow `Normal` extrinsics to fill up the block up to 75%, the rest can be used
+/// by  Operational  extrinsics.
+pub const NORMAL_DISPATCH_RATIO: Perbill = Perbill::from_percent(75);
 
 parameter_types! {
 	/// NFT metadata limit
