+++ runtime/basilisk/src/lib.rs
@@ -1,6 +1,6 @@
 // This file is part of Basilisk-node.
 
-// Copyright (C) 2020-2022  Intergalactic, Limited (GIB).
+// Copyright (C) 2020-2023  Intergalactic, Limited (GIB).
 // SPDX-License-Identifier: Apache-2.0
 
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -24,67 +24,46 @@
 #![allow(clippy::from_over_into)]
 #![allow(clippy::match_like_matches_macro)]
 
-#[cfg(test)]
-mod tests;
-
 // Make the WASM binary available.
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
-use frame_system::{EnsureRoot, RawOrigin};
-use hydradx_adapters::inspect::MultiInspectAdapter;
-use sp_api::impl_runtime_apis;
-use sp_core::{ConstU32, OpaqueMetadata};
-use sp_runtime::{
-	app_crypto::sp_core::crypto::UncheckedFrom,
+#[cfg(test)]
+mod tests;
+
+mod benchmarking;
+mod migrations;
+pub mod weights;
+
+mod adapter;
+mod democracy;
+mod finance;
+mod system;
+mod xcm;
+
+pub use democracy::*;
+pub use finance::*;
+pub use system::*;
+pub use xcm::*;
+
+pub use primitives::{AccountId, Amount, AssetId, Balance, BlockNumber, CollectionId, Hash, Index, ItemId, Signature};
+
+use frame_support::sp_runtime::{
 	create_runtime_str, generic, impl_opaque_keys,
-	traits::{AccountIdConversion, BlakeTwo256, Block as BlockT, IdentityLookup},
+	traits::{AccountIdConversion, BlakeTwo256, Block as BlockT},
 	transaction_validity::{TransactionSource, TransactionValidity},
-	ApplyExtrinsicResult, Perbill,
+	ApplyExtrinsicResult,
 };
-use sp_std::convert::From;
-use sp_std::marker::PhantomData;
-use sp_std::prelude::*;
+use sp_api::impl_runtime_apis;
+pub use sp_consensus_aura::sr25519::AuthorityId as AuraId;
+use sp_core::{ConstU32, OpaqueMetadata};
+use sp_std::{convert::From, marker::PhantomData, prelude::*, vec};
 #[cfg(feature = "std")]
 use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
 
-#[cfg(feature = "runtime-benchmarks")]
-use codec::Decode;
-
 // A few exports that help ease life for downstream crates.
-use frame_support::{
-	construct_runtime,
-	dispatch::DispatchClass,
-	parameter_types,
-	traits::{
-		AsEnsureOriginWithArg, Contains, Defensive, EitherOfDiverse, EnsureOrigin, EqualPrivilegeOnly, Get,
-		InstanceFilter, NeverEnsureOrigin, U128CurrencyToVote,
-	},
-	weights::{
-		constants::{BlockExecutionWeight, RocksDbWeight},
-		ConstantMultiplier, Weight, WeightToFeeCoefficient, WeightToFeeCoefficients, WeightToFeePolynomial,
-	},
-	BoundedVec,
-};
-use hydradx_traits::{AssetPairAccountIdFor, OraclePeriod};
-use pallet_transaction_payment::TargetedFeeAdjustment;
-pub use sp_consensus_aura::sr25519::AuthorityId as AuraId;
-
-mod xcm;
-
-mod benchmarking;
-
-use orml_tokens::CurrencyAdapter;
-use orml_traits::currency::MutationHooks;
-use pallet_collective::EnsureProportionAtLeast;
-use pallet_currencies::BasicCurrencyAdapter;
-
-use common_runtime::locked_balance::MultiCurrencyLockedBalance;
-pub use common_runtime::*;
-use pallet_transaction_multi_payment::{AddTxAssetOnAccount, DepositAll, RemoveTxAssetOnKilled, TransferFees};
-
-mod migrations;
+use frame_support::{construct_runtime, weights::Weight};
 
 /// Opaque types. These are used by the CLI to instantiate machinery that don't need to know
 /// the specifics of the runtime. They can then be made to be agnostic over specific formats
@@ -113,7 +92,7 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
 	spec_name: create_runtime_str!("basilisk"),
 	impl_name: create_runtime_str!("basilisk"),
 	authoring_version: 1,
-	spec_version: 99,
+	spec_version: 100,
 	impl_version: 0,
 	apis: RUNTIME_API_VERSIONS,
 	transaction_version: 1,
@@ -129,68 +108,20 @@ pub fn native_version() -> NativeVersion {
 	}
 }
 
-pub struct BaseFilter;
-impl Contains<RuntimeCall> for BaseFilter {
-	fn contains(call: &RuntimeCall) -> bool {
-		if matches!(
-			call,
-			RuntimeCall::System(_) | RuntimeCall::Timestamp(_) | RuntimeCall::ParachainSystem(_)
-		) {
-			// always allow
-			// Note: this is done to avoid unnecessary check of paused storage.
-			return true;
-		}
-
-		if pallet_transaction_pause::PausedTransactionFilter::<Runtime>::contains(call) {
-			// if paused, dont allow!
-			return false;
-		}
-
-		match call {
-			RuntimeCall::Uniques(_) => false,
-			RuntimeCall::PolkadotXcm(_) => false,
-			RuntimeCall::OrmlXcm(_) => false,
-			_ => true,
-		}
-	}
+pub fn get_all_module_accounts() -> vec::Vec<AccountId> {
+	vec![
+		system::TreasuryPalletId::get().into_account_truncating(),
+		finance::VestingPalletId::get().into_account_truncating(),
+	]
 }
 
-use common_runtime::adapter::OrmlTokensAdapter;
-use primitives::{CollectionId, ItemId};
-use smallvec::smallvec;
 use sp_runtime::traits::BlockNumberProvider;
-
-pub struct WeightToFee;
-impl WeightToFeePolynomial for WeightToFee {
-	type Balance = Balance;
-
-	/// Handles converting a weight scalar to a fee value, based on the scale and granularity of the
-	/// node's balance type.
-	///
-	/// This should typically create a mapping between the following ranges:
-	///   - [0, MAXIMUM_BLOCK_WEIGHT]
-	///   - [Balance::min, Balance::max]
-	///
-	/// Yet, it can be used for any other sort of change to weight-fee. Some examples being:
-	///   - Setting it to `0` will essentially disable the weight fee.
-	///   - Setting it to `1` will cause the literal `#[weight = x]` values to be charged.
-	fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-		let p = 11 * CENTS;
-		let q = Balance::from(ExtrinsicBaseWeight::get().ref_time());
-		smallvec![WeightToFeeCoefficient {
-			degree: 1,
-			negative: false,
-			coeff_frac: Perbill::from_rational(p % q, q),
-			coeff_integer: p / q,
-		}]
-	}
-}
 
 // Relay chain Block number provider.
 // Reason why the implementation is different for benchmarks is that it is not possible
 // to set or change the block number in a benchmark using parachain system pallet.
 // That's why we revert to using the system pallet in the benchmark.
-pub struct RelayChainBlockNumberProvider<T>(sp_std::marker::PhantomData<T>);
+pub struct RelayChainBlockNumberProvider<T>(PhantomData<T>);
 
 #[cfg(not(feature = "runtime-benchmarks"))]
 impl<T: cumulus_pallet_parachain_system::Config + orml_tokens::Config> BlockNumberProvider
@@ -216,772 +147,6 @@ impl<T: frame_system::Config> BlockNumberProvider for RelayChainBlockNumberProvi
 	fn current_block_number() -> Self::BlockNumber {
 		frame_system::Pallet::<T>::current_block_number()
 	}
-}
-
-pub struct AssetPairAccountId<T: frame_system::Config>(PhantomData<T>);
-impl<T: frame_system::Config> AssetPairAccountIdFor<AssetId, T::AccountId> for AssetPairAccountId<T>
-where
-	T::AccountId: UncheckedFrom<T::Hash> + AsRef<[u8]>,
-{
-	fn from_assets(asset_a: AssetId, asset_b: AssetId, identifier: &str) -> T::AccountId {
-		let mut buf: Vec<u8> = identifier.as_bytes().to_vec();
-
-		if asset_a < asset_b {
-			buf.extend_from_slice(&asset_a.to_le_bytes());
-			buf.extend_from_slice(&asset_b.to_le_bytes());
-		} else {
-			buf.extend_from_slice(&asset_b.to_le_bytes());
-			buf.extend_from_slice(&asset_a.to_le_bytes());
-		}
-		T::AccountId::unchecked_from(<T::Hashing as frame_support::sp_runtime::traits::Hash>::hash(&buf[..]))
-	}
-}
-
-parameter_types! {
-	pub const Version: RuntimeVersion = VERSION;
-	/// Block weights base values and limits.
-	pub BlockWeights: frame_system::limits::BlockWeights = frame_system::limits::BlockWeights::builder()
-		.base_block(BlockExecutionWeight::get())
-		.for_class(DispatchClass::all(), |weights| {
-			weights.base_extrinsic = ExtrinsicBaseWeight::get();
-		})
-		.for_class(DispatchClass::Normal, |weights| {
-			weights.max_total = Some(NORMAL_DISPATCH_RATIO * MAXIMUM_BLOCK_WEIGHT);
-		})
-		.for_class(DispatchClass::Operational, |weights| {
-			weights.max_total = Some(MAXIMUM_BLOCK_WEIGHT);
-			// Operational transactions have an extra reserved space, so that they
-			// are included even if block reachd `MAXIMUM_BLOCK_WEIGHT`.
-			weights.reserved = Some(
-				MAXIMUM_BLOCK_WEIGHT - NORMAL_DISPATCH_RATIO * MAXIMUM_BLOCK_WEIGHT,
-			);
-		})
-		.avg_block_initialization(AVERAGE_ON_INITIALIZE_RATIO)
-		.build_or_panic();
-
-	pub ExtrinsicBaseWeight: Weight = common_runtime::BasiliskExtrinsicBaseWeight::get();
-}
-
-// Configure FRAME pallets to include in runtime.
-
-impl frame_system::Config for Runtime {
-	/// The basic call filter to use in dispatchable.
-	type BaseCallFilter = BaseFilter;
-	type BlockWeights = BlockWeights;
-	type BlockLength = BlockLength;
-	/// The ubiquitous origin type.
-	type RuntimeOrigin = RuntimeOrigin;
-	/// The aggregated dispatch type that is available for extrinsics.
-	type RuntimeCall = RuntimeCall;
-	/// The index type for storing how many extrinsics an account has signed.
-	type Index = Index;
-	/// The index type for blocks.
-	type BlockNumber = BlockNumber;
-	/// The type for hashing blocks and tries.
-	type Hash = Hash;
-	/// The hashing algorithm used.
-	type Hashing = BlakeTwo256;
-	/// The identifier used to distinguish between accounts.
-	type AccountId = AccountId;
-	/// The lookup mechanism to get account ID from whatever is passed in dispatchers.
-	type Lookup = IdentityLookup<AccountId>;
-	/// The header type.
-	type Header = generic::Header<BlockNumber, BlakeTwo256>;
-	/// The ubiquitous event type.
-	type RuntimeEvent = RuntimeEvent;
-	/// Maximum number of block number to block hash mappings to keep (oldest pruned first).
-	type BlockHashCount = BlockHashCount;
-	/// The weight of database operations that the runtime can invoke.
-	type DbWeight = RocksDbWeight;
-	/// The weight of the overhead invoked on the block import process, independent of the
-	/// extrinsics included in that block.
-	/// Version of the runtime.
-	type Version = Version;
-	/// Converts a module to the index of the module in `construct_runtime!`.
-	///
-	/// This type is being generated by `construct_runtime!`.
-	type PalletInfo = PalletInfo;
-	/// The data to be stored in an account.
-	type AccountData = pallet_balances::AccountData<Balance>;
-	/// What to do if a new account is created.
-	type OnNewAccount = ();
-	/// What to do if an account is fully reaped from the system.
-	type OnKilledAccount = ();
-	/// Weight information for the extrinsics of this pallet.
-	type SystemWeightInfo = weights::system::BasiliskWeight<Runtime>;
-	type SS58Prefix = SS58Prefix;
-	type OnSetCode = cumulus_pallet_parachain_system::ParachainSetCode<Self>;
-	type MaxConsumers = frame_support::traits::ConstU32<16>;
-}
-
-impl pallet_timestamp::Config for Runtime {
-	/// A timestamp: milliseconds since the unix epoch.
-	type Moment = u64;
-	type OnTimestampSet = ();
-	type MinimumPeriod = MinimumPeriod;
-	type WeightInfo = weights::timestamp::BasiliskWeight<Runtime>;
-}
-
-impl pallet_balances::Config for Runtime {
-	type MaxLocks = MaxLocks;
-	/// The type for recording an account's balance.
-	type Balance = Balance;
-	/// The ubiquitous event type.
-	type RuntimeEvent = RuntimeEvent;
-	type DustRemoval = Treasury;
-	type ExistentialDeposit = NativeExistentialDeposit;
-	type AccountStore = System;
-	type WeightInfo = weights::balances::BasiliskWeight<Runtime>;
-	type MaxReserves = MaxReserves;
-	type ReserveIdentifier = ();
-}
-
-/// Parameterized slow adjusting fee updated based on
-/// https://w3f-research.readthedocs.io/en/latest/polkadot/overview/2-token-economics.html?highlight=token%20economics#-2.-slow-adjusting-mechanism
-pub type SlowAdjustingFeeUpdate<R> =
-	TargetedFeeAdjustment<R, TargetBlockFullness, AdjustmentVariable, MinimumMultiplier, MaximumMultiplier>;
-
-impl pallet_transaction_payment::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type OnChargeTransaction = TransferFees<Currencies, DepositAll<Runtime>, TreasuryAccount>;
-	type OperationalFeeMultiplier = ();
-	type WeightToFee = WeightToFee;
-	type LengthToFee = ConstantMultiplier<Balance, TransactionByteFee>;
-	type FeeMultiplierUpdate = SlowAdjustingFeeUpdate<Self>;
-}
-
-parameter_types! {
-	pub TreasuryAccount: AccountId = Treasury::account_id();
-}
-
-impl pallet_transaction_multi_payment::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type AcceptedCurrencyOrigin = MajorityTechCommitteeOrRoot;
-	type Currencies = Currencies;
-	type SpotPriceProvider = pallet_xyk::XYKSpotPrice<Runtime>;
-	type WeightInfo = weights::payment::BasiliskWeight<Runtime>;
-	type WeightToFee = WeightToFee;
-	type NativeAssetId = NativeAssetId;
-}
-
-impl InstanceFilter<RuntimeCall> for ProxyType {
-	fn filter(&self, c: &RuntimeCall) -> bool {
-		match self {
-			ProxyType::Any => true,
-			ProxyType::CancelProxy => matches!(c, RuntimeCall::Proxy(pallet_proxy::Call::reject_announcement { .. })),
-			ProxyType::Governance => matches!(
-				c,
-				RuntimeCall::Democracy(..)
-					| RuntimeCall::Council(..)
-					| RuntimeCall::TechnicalCommittee(..)
-					| RuntimeCall::Elections(..)
-					| RuntimeCall::Treasury(..)
-					| RuntimeCall::Tips(..)
-					| RuntimeCall::Utility(..)
-			),
-			ProxyType::Exchange => matches!(c, RuntimeCall::XYK(..) | RuntimeCall::LBP(..) | RuntimeCall::NFT(..)),
-			// Transfer group doesn't include cross-chain transfers
-			ProxyType::Transfer => matches!(
-				c,
-				RuntimeCall::Balances(..) | RuntimeCall::Currencies(..) | RuntimeCall::Tokens(..)
-			),
-		}
-	}
-	fn is_superset(&self, o: &Self) -> bool {
-		match (self, o) {
-			(x, y) if x == y => true,
-			(ProxyType::Any, _) => true,
-			(_, ProxyType::Any) => false,
-			_ => false,
-		}
-	}
-}
-
-impl pallet_proxy::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type RuntimeCall = RuntimeCall;
-	type Currency = Balances;
-	type ProxyType = ProxyType;
-	type ProxyDepositBase = ProxyDepositBase;
-	type ProxyDepositFactor = ProxyDepositFactor;
-	type MaxProxies = MaxProxies;
-	type WeightInfo = ();
-	type MaxPending = MaxPending;
-	type CallHasher = BlakeTwo256;
-	type AnnouncementDepositBase = AnnouncementDepositBase;
-	type AnnouncementDepositFactor = AnnouncementDepositFactor;
-}
-
-pub struct DustRemovalWhitelist;
-impl Contains<AccountId> for DustRemovalWhitelist {
-	fn contains(a: &AccountId) -> bool {
-		get_all_module_accounts().contains(a) || pallet_duster::DusterWhitelist::<Runtime>::contains(a)
-	}
-}
-
-pub struct CurrencyHooks;
-impl MutationHooks<AccountId, AssetId, Balance> for CurrencyHooks {
-	type OnDust = Duster;
-	type OnSlash = ();
-	type PreDeposit = ();
-	type PostDeposit = ();
-	type PreTransfer = ();
-	type PostTransfer = ();
-	type OnNewTokenAccount = AddTxAssetOnAccount<Runtime>;
-	type OnKilledTokenAccount = RemoveTxAssetOnKilled<Runtime>;
-}
-
-/// Tokens Configurations
-impl orml_tokens::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Balance = Balance;
-	type Amount = Amount;
-	type CurrencyId = AssetId;
-	type WeightInfo = weights::tokens::BasiliskWeight<Runtime>;
-	type ExistentialDeposits = AssetRegistry;
-	type MaxLocks = MaxLocks;
-	type MaxReserves = MaxReserves;
-	type ReserveIdentifier = ();
-	type DustRemovalWhitelist = DustRemovalWhitelist;
-	type CurrencyHooks = CurrencyHooks;
-}
-
-// The latest versions of the orml-currencies pallet don't emit events.
-// The infrastructure relies on the events from this pallet, so we use the latest version of
-// the pallet that contains and emit events and was updated to the polkadot version we use.
-impl pallet_currencies::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type MultiCurrency = OrmlTokensAdapter<Runtime>;
-	type NativeCurrency = BasicCurrencyAdapter<Runtime, Balances, Amount, BlockNumber>;
-	type GetNativeCurrencyId = NativeAssetId;
-	type WeightInfo = weights::currencies::BasiliskWeight<Runtime>;
-}
-
-impl pallet_duster::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Balance = Balance;
-	type Amount = Amount;
-	type CurrencyId = AssetId;
-	type MultiCurrency = Currencies;
-	type MinCurrencyDeposits = AssetRegistry;
-	type Reward = DustingReward;
-	type NativeCurrencyId = NativeAssetId;
-	type BlacklistUpdateOrigin = MajorityTechCommitteeOrRoot;
-	type WeightInfo = weights::duster::BasiliskWeight<Runtime>;
-}
-
-impl pallet_asset_registry::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type RegistryOrigin = SuperMajorityTechCommitteeOrRoot;
-	type AssetId = AssetId;
-	type Balance = Balance;
-	type AssetNativeLocation = AssetLocation;
-	type StringLimit = RegistryStrLimit;
-	type SequentialIdStartAt = SequentialIdOffset;
-	type NativeAssetId = NativeAssetId;
-	type WeightInfo = weights::asset_registry::BasiliskWeight<Runtime>;
-}
-
-impl pallet_xyk::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type AssetRegistry = AssetRegistry;
-	type AssetPairAccountId = AssetPairAccountId<Self>;
-	type Currency = Currencies;
-	type NativeAssetId = NativeAssetId;
-	type WeightInfo = weights::xyk::BasiliskWeight<Runtime>;
-	type GetExchangeFee = ExchangeFee;
-	type MinTradingLimit = MinTradingLimit;
-	type MinPoolLiquidity = MinPoolLiquidity;
-	type MaxInRatio = MaxInRatio;
-	type MaxOutRatio = MaxOutRatio;
-	type CanCreatePool = pallet_lbp::DisallowWhenLBPPoolRunning<Runtime>;
-	type AMMHandler = pallet_ema_oracle::OnActivityHandler<Runtime>;
-	type DiscountedFee = DiscountedFee;
-	type NonDustableWhitelistHandler = Duster;
-}
-
-impl pallet_lbp::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type MultiCurrency = Currencies;
-	type LockedBalance = MultiCurrencyLockedBalance<Runtime>;
-	type CreatePoolOrigin = SuperMajorityTechCommitteeOrRoot;
-	type LBPWeightFunction = pallet_lbp::LBPWeightFunction;
-	type AssetPairAccountId = AssetPairAccountId<Self>;
-	type MinTradingLimit = MinTradingLimit;
-	type MinPoolLiquidity = MinPoolLiquidity;
-	type MaxInRatio = MaxInRatio;
-	type MaxOutRatio = MaxOutRatio;
-	type WeightInfo = weights::lbp::BasiliskWeight<Runtime>;
-	type BlockNumberProvider = RelayChainBlockNumberProvider<Runtime>;
-}
-
-// Parachain Config
-
-parameter_types! {
-	pub ReservedXcmpWeight: Weight = BlockWeights::get().max_block / 4;
-	pub ReservedDmpWeight: Weight = BlockWeights::get().max_block / 4;
-}
-
-impl cumulus_pallet_parachain_system::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type OnSystemEvent = pallet_relaychain_info::OnValidationDataHandler<Runtime>;
-	type SelfParaId = ParachainInfo;
-	type OutboundXcmpMessageSource = XcmpQueue;
-	type DmpMessageHandler = DmpQueue;
-	type ReservedDmpWeight = ReservedDmpWeight;
-	type XcmpMessageHandler = XcmpQueue;
-	type ReservedXcmpWeight = ReservedXcmpWeight;
-	type CheckAssociatedRelayNumber = cumulus_pallet_parachain_system::RelayNumberStrictlyIncreases;
-}
-
-impl pallet_aura::Config for Runtime {
-	type AuthorityId = AuraId;
-	type MaxAuthorities = MaxAuthorities;
-	type DisabledValidators = ();
-}
-
-impl parachain_info::Config for Runtime {}
-
-impl cumulus_pallet_aura_ext::Config for Runtime {}
-
-parameter_types! {
-	pub ReserveCollectionIdUpTo: u128 = 999_999;
-}
-
-impl pallet_nft::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type WeightInfo = weights::nft::BasiliskWeight<Runtime>;
-	type NftCollectionId = CollectionId;
-	type NftItemId = ItemId;
-	type CollectionType = pallet_nft::CollectionType;
-	type Permissions = pallet_nft::NftPermissions;
-	type ReserveCollectionIdUpTo = ReserveCollectionIdUpTo;
-}
-
-type MajorityCouncilOrRoot =
-	EitherOfDiverse<EnsureProportionAtLeast<AccountId, CouncilCollective, 1, 2>, EnsureRoot<AccountId>>;
-type UnanimousCouncilOrRoot =
-	EitherOfDiverse<EnsureProportionAtLeast<AccountId, CouncilCollective, 1, 1>, EnsureRoot<AccountId>>;
-type SuperMajorityCouncilOrRoot =
-	EitherOfDiverse<EnsureProportionAtLeast<AccountId, CouncilCollective, 2, 3>, EnsureRoot<AccountId>>;
-type SuperMajorityTechCommitteeOrRoot =
-	EitherOfDiverse<EnsureProportionAtLeast<AccountId, TechnicalCollective, 2, 3>, EnsureRoot<AccountId>>;
-type UnanimousTechCommitteeOrRoot =
-	EitherOfDiverse<EnsureProportionAtLeast<AccountId, TechnicalCollective, 1, 1>, EnsureRoot<AccountId>>;
-type MajorityTechCommitteeOrRoot =
-	EitherOfDiverse<EnsureProportionAtLeast<AccountId, TechnicalCollective, 1, 2>, EnsureRoot<AccountId>>;
-
-impl pallet_democracy::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type EnactmentPeriod = EnactmentPeriod;
-	type LaunchPeriod = LaunchPeriod;
-	type VotingPeriod = VotingPeriod;
-	type MinimumDeposit = MinimumDeposit;
-	/// A straight majority of the council can decide what their next motion is.
-	type ExternalOrigin = MajorityCouncilOrRoot;
-	/// A majority can have the next scheduled referendum be a straight majority-carries vote
-	type ExternalMajorityOrigin = MajorityCouncilOrRoot;
-	/// A unanimous council can have the next scheduled referendum be a straight default-carries
-	/// (NTB) vote.
-	type ExternalDefaultOrigin = UnanimousCouncilOrRoot;
-	/// Two thirds of the technical committee can have an ExternalMajority/ExternalDefault vote
-	/// be tabled immediately and with a shorter voting/enactment period.
-	type FastTrackOrigin = MajorityTechCommitteeOrRoot;
-	type InstantOrigin = UnanimousTechCommitteeOrRoot;
-	type InstantAllowed = InstantAllowed;
-	type FastTrackVotingPeriod = FastTrackVotingPeriod;
-	// To cancel a proposal which has been passed, 2/3 of the council must agree to it.
-	type CancellationOrigin = SuperMajorityCouncilOrRoot;
-	// To cancel a proposal before it has been passed, the technical committee must be unanimous or
-	// Root must agree.
-	type CancelProposalOrigin = UnanimousTechCommitteeOrRoot;
-	type BlacklistOrigin = EnsureRoot<AccountId>;
-	// Any single technical committee member may veto a coming council proposal, however they can
-	// only do it once and it lasts only for the cooloff period.
-	type VetoOrigin = pallet_collective::EnsureMember<AccountId, TechnicalCollective>;
-	type CooloffPeriod = CooloffPeriod;
-	type Slash = Treasury;
-	type Scheduler = Scheduler;
-	type PalletsOrigin = OriginCaller;
-	type MaxVotes = MaxVotes;
-	type WeightInfo = weights::democracy::BasiliskWeight<Runtime>;
-	type MaxProposals = MaxProposals;
-	type Preimages = Preimage;
-	type MaxDeposits = ConstU32<100>;
-	type MaxBlacklisted = ConstU32<100>;
-	type VoteLockingPeriod = VoteLockingPeriod;
-}
-
-impl pallet_elections_phragmen::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type PalletId = ElectionsPhragmenPalletId;
-	type Currency = Balances;
-	type ChangeMembers = Council;
-	type InitializeMembers = (); // Set to () if defined in chain spec
-	type CurrencyToVote = U128CurrencyToVote;
-	type CandidacyBond = CandidacyBond;
-	type VotingBondBase = VotingBondBase;
-	type VotingBondFactor = VotingBondFactor;
-	type LoserCandidate = Treasury;
-	type KickedMember = Treasury;
-	type DesiredMembers = DesiredMembers;
-	type DesiredRunnersUp = DesiredRunnersUp;
-	type TermDuration = TermDuration;
-	type MaxCandidates = MaxElectionCandidates;
-	type MaxVoters = MaxElectionVoters;
-	type WeightInfo = ();
-}
-
-type CouncilCollective = pallet_collective::Instance1;
-impl pallet_collective::Config<CouncilCollective> for Runtime {
-	type RuntimeOrigin = RuntimeOrigin;
-	type Proposal = RuntimeCall;
-	type RuntimeEvent = RuntimeEvent;
-	type MotionDuration = CouncilMotionDuration;
-	type MaxProposals = CouncilMaxProposals;
-	type MaxMembers = CouncilMaxMembers;
-	type DefaultVote = pallet_collective::PrimeDefaultVote;
-	type WeightInfo = ();
-}
-
-type TechnicalCollective = pallet_collective::Instance2;
-impl pallet_collective::Config<TechnicalCollective> for Runtime {
-	type RuntimeOrigin = RuntimeOrigin;
-	type Proposal = RuntimeCall;
-	type RuntimeEvent = RuntimeEvent;
-	type MotionDuration = TechnicalMotionDuration;
-	type MaxProposals = TechnicalMaxProposals;
-	type MaxMembers = TechnicalMaxMembers;
-	type DefaultVote = pallet_collective::PrimeDefaultVote;
-	type WeightInfo = ();
-}
-
-impl pallet_treasury::Config for Runtime {
-	type PalletId = TreasuryPalletId;
-	type Currency = Balances;
-	type ApproveOrigin = SuperMajorityCouncilOrRoot;
-	type RejectOrigin = MajorityCouncilOrRoot;
-	type RuntimeEvent = RuntimeEvent;
-	type OnSlash = Treasury;
-	type ProposalBond = ProposalBond;
-	type ProposalBondMinimum = ProposalBondMinimum;
-	type ProposalBondMaximum = ProposalBondMaximum;
-	type SpendPeriod = SpendPeriod;
-	type Burn = Burn;
-	type BurnDestination = ();
-	type WeightInfo = weights::treasury::BasiliskWeight<Runtime>;
-	type SpendFunds = ();
-	type MaxApprovals = MaxApprovals;
-	type SpendOrigin = NeverEnsureOrigin<Balance>;
-}
-
-parameter_types! {
-	pub MaximumSchedulerWeight: Weight = Perbill::from_percent(10) * BlockWeights::get().max_block;
-	pub const MaxScheduledPerBlock: u32 = 50;
-	pub const NoPreimagePostponement: Option<u32> = Some(5 * MINUTES);
-}
-
-impl pallet_scheduler::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type RuntimeOrigin = RuntimeOrigin;
-	type PalletsOrigin = OriginCaller;
-	type RuntimeCall = RuntimeCall;
-	type MaximumWeight = MaximumSchedulerWeight;
-	type ScheduleOrigin = EnsureRoot<AccountId>;
-	type MaxScheduledPerBlock = MaxScheduledPerBlock;
-	type WeightInfo = weights::scheduler::BasiliskWeight<Runtime>;
-	type OriginPrivilegeCmp = EqualPrivilegeOnly;
-	type Preimages = Preimage;
-}
-
-impl pallet_utility::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type RuntimeCall = RuntimeCall;
-	type WeightInfo = weights::utility::BasiliskWeight<Runtime>;
-	type PalletsOrigin = OriginCaller;
-}
-
-impl pallet_authorship::Config for Runtime {
-	type FindAuthor = pallet_session::FindAccountFromAuthorIndex<Self, Aura>;
-	type EventHandler = (CollatorSelection,);
-}
-
-impl pallet_tips::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type DataDepositPerByte = DataDepositPerByte;
-	type MaximumReasonLength = MaximumReasonLength;
-	type Tippers = Elections;
-	type TipCountdown = TipCountdown;
-	type TipFindersFee = TipFindersFee;
-	type TipReportDepositBase = TipReportDepositBase;
-	type WeightInfo = ();
-}
-
-impl pallet_collator_selection::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type UpdateOrigin = MajorityTechCommitteeOrRoot;
-	type PotId = PotId;
-	type MaxCandidates = MaxCandidates;
-	type MinCandidates = MinCandidates;
-	type MaxInvulnerables = MaxInvulnerables;
-	// should be a multiple of session or things will get inconsistent
-	type KickThreshold = Period;
-	type ValidatorId = <Self as frame_system::Config>::AccountId;
-	type ValidatorIdOf = pallet_collator_selection::IdentityCollator;
-	type ValidatorRegistration = Session;
-	type WeightInfo = weights::collator_selection::BasiliskWeight<Runtime>;
-}
-
-impl pallet_session::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type ValidatorId = <Self as frame_system::Config>::AccountId;
-	// we don't have stash and controller, thus we don't need the convert as well.
-	type ValidatorIdOf = pallet_collator_selection::IdentityCollator;
-	type ShouldEndSession = pallet_session::PeriodicSessions<Period, Offset>;
-	type NextSessionRotation = pallet_session::PeriodicSessions<Period, Offset>;
-	type SessionManager = CollatorRewards;
-	// Essentially just Aura, but lets be pedantic.
-	type SessionHandler = <opaque::SessionKeys as sp_runtime::traits::OpaqueKeys>::KeyTypeIdProviders;
-	type Keys = opaque::SessionKeys;
-	type WeightInfo = ();
-}
-
-pub struct RootAsVestingPallet;
-impl EnsureOrigin<RuntimeOrigin> for RootAsVestingPallet {
-	type Success = AccountId;
-
-	fn try_origin(o: RuntimeOrigin) -> Result<Self::Success, RuntimeOrigin> {
-		Into::<Result<RawOrigin<AccountId>, RuntimeOrigin>>::into(o).and_then(|o| match o {
-			RawOrigin::Root => Ok(VestingPalletId::get().into_account_truncating()),
-			r => Err(RuntimeOrigin::from(r)),
-		})
-	}
-
-	#[cfg(feature = "runtime-benchmarks")]
-	fn try_successful_origin() -> Result<RuntimeOrigin, ()> {
-		let zero_account_id = AccountId::decode(&mut sp_runtime::traits::TrailingZeroInput::zeroes())
-			.expect("infinite length input; no invalid inputs for type; qed");
-		Ok(RuntimeOrigin::from(RawOrigin::Signed(zero_account_id)))
-	}
-}
-
-impl orml_vesting::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type MinVestedTransfer = MinVestedTransfer;
-	type VestedTransferOrigin = RootAsVestingPallet;
-	type WeightInfo = weights::vesting::BasiliskWeight<Runtime>;
-	type MaxVestingSchedules = MaxVestingSchedules;
-	type BlockNumberProvider = RelayChainBlockNumberProvider<Runtime>;
-}
-
-parameter_types! {
-	pub const MinimumOfferAmount: Balance = UNITS / 100;
-	pub const RoyaltyBondAmount: Balance = 0;
-}
-
-pub struct RelayChainAssetId;
-impl Get<AssetId> for RelayChainAssetId {
-	fn get() -> AssetId {
-		let invalid_id = pallet_asset_registry::Pallet::<Runtime>::next_asset_id().defensive_unwrap_or(AssetId::MAX);
-
-		match pallet_asset_registry::Pallet::<Runtime>::location_to_asset(RELAY_CHAIN_ASSET_LOCATION) {
-			Some(asset_id) => asset_id,
-			None => invalid_id,
-		}
-	}
-}
-
-type KusamaCurrency = CurrencyAdapter<Runtime, RelayChainAssetId>;
-
-impl pallet_marketplace::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = KusamaCurrency;
-	type WeightInfo = pallet_marketplace::weights::BasiliskWeight<Runtime>;
-	type MinimumOfferAmount = MinimumOfferAmount;
-	type RoyaltyBondAmount = RoyaltyBondAmount;
-}
-
-pub mod ksm {
-	use primitives::Balance;
-
-	pub const UNITS: Balance = 1_000_000_000_000;
-	pub const CENTS: Balance = UNITS / 30_000;
-	pub const MILLICENTS: Balance = CENTS / 1_000;
-
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		(items as Balance * 2_000 * CENTS + (bytes as Balance) * 100 * MILLICENTS) / 10
-	}
-}
-
-parameter_types! {
-	pub const CollectionDeposit: Balance = 0;
-	pub const ItemDeposit: Balance = 0;
-	pub const KeyLimit: u32 = 256;	// Max 256 bytes per key
-	pub const ValueLimit: u32 = 1024;	// Max 1024 bytes per value
-	pub const UniquesMetadataDepositBase: Balance = ksm::deposit(1,129);
-	pub const AttributeDepositBase: Balance = ksm::deposit(1,0);
-	pub const DepositPerByte: Balance = ksm::deposit(0,1);
-}
-
-impl pallet_uniques::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type CollectionId = CollectionId;
-	type ItemId = ItemId;
-	type Currency = KusamaCurrency;
-	type ForceOrigin = SuperMajorityCouncilOrRoot;
-	// Standard collection creation is disallowed
-	type CreateOrigin = AsEnsureOriginWithArg<NeverEnsureOrigin<AccountId>>;
-	type Locker = ();
-	type CollectionDeposit = CollectionDeposit;
-	type ItemDeposit = ItemDeposit;
-	type MetadataDepositBase = UniquesMetadataDepositBase;
-	type AttributeDepositBase = AttributeDepositBase;
-	type DepositPerByte = DepositPerByte;
-	type StringLimit = primitives::UniquesStringLimit;
-	type KeyLimit = KeyLimit;
-	type ValueLimit = ValueLimit;
-	type WeightInfo = ();
-	#[cfg(feature = "runtime-benchmarks")]
-	type Helper = ();
-}
-
-impl pallet_relaychain_info::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type RelaychainBlockNumberProvider = RelayChainBlockNumberProvider<Runtime>;
-}
-
-impl pallet_xyk_liquidity_mining::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type MultiCurrency = Currencies;
-	type CreateOrigin = UnanimousTechCommitteeOrRoot;
-	type PalletId = LMPalletId;
-	type NftCollectionId = LiquidityMiningNftCollectionId;
-	type AMM = XYK;
-	type WeightInfo = weights::xyk_liquidity_mining::BasiliskWeight<Runtime>;
-	type NFTHandler = NFT;
-	type LiquidityMiningHandler = XYKWarehouseLM;
-	type NonDustableWhitelistHandler = Duster;
-}
-
-type XYKLiquidityMiningInstance = warehouse_liquidity_mining::Instance1;
-impl warehouse_liquidity_mining::Config<XYKLiquidityMiningInstance> for Runtime {
-	type AssetId = AssetId;
-	type MultiCurrency = Currencies;
-	type PalletId = WarehouseLMPalletId;
-	type MinTotalFarmRewards = MinTotalFarmRewards;
-	type MinPlannedYieldingPeriods = MinPlannedYieldingPeriods;
-	type BlockNumberProvider = RelayChainBlockNumberProvider<Runtime>;
-	type AmmPoolId = AccountId;
-	type MaxFarmEntriesPerDeposit = MaxEntriesPerDeposit;
-	type MaxYieldFarmsPerGlobalFarm = MaxYieldFarmsPerGlobalFarm;
-	type AssetRegistry = AssetRegistry;
-	type NonDustableWhitelistHandler = Duster;
-	type RuntimeEvent = RuntimeEvent;
-	type PriceAdjustment = warehouse_liquidity_mining::DefaultPriceAdjustment;
-}
-
-parameter_types! {
-	pub const PreimageMaxSize: u32 = 4096 * 1024;
-	pub PreimageBaseDeposit: Balance = deposit(2, 64);
-	pub PreimageByteDeposit: Balance = deposit(0, 1);
-}
-
-impl pallet_preimage::Config for Runtime {
-	type WeightInfo = ();
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type ManagerOrigin = EnsureRoot<AccountId>;
-	type BaseDeposit = PreimageBaseDeposit;
-	type ByteDeposit = PreimageByteDeposit;
-}
-
-impl pallet_identity::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type BasicDeposit = BasicDeposit;
-	type FieldDeposit = FieldDeposit;
-	type SubAccountDeposit = SubAccountDeposit;
-	type MaxSubAccounts = MaxSubAccounts;
-	type MaxAdditionalFields = MaxAdditionalFields;
-	type MaxRegistrars = MaxRegistrars;
-	type Slashed = Treasury;
-	type ForceOrigin = MajorityCouncilOrRoot;
-	type RegistrarOrigin = MajorityCouncilOrRoot;
-	type WeightInfo = ();
-}
-
-impl pallet_multisig::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type RuntimeCall = RuntimeCall;
-	type Currency = Balances;
-	type DepositBase = DepositBase;
-	type DepositFactor = DepositFactor;
-	type MaxSignatories = MaxSignatories;
-	type WeightInfo = ();
-}
-
-impl pallet_transaction_pause::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type UpdateOrigin = MajorityTechCommitteeOrRoot;
-	type WeightInfo = weights::transaction_pause::BasiliskWeight<Runtime>;
-}
-parameter_types! {
-	pub const MaxNumberOfTrades: u8 = 5;
-}
-
-impl pallet_route_executor::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type AssetId = AssetId;
-	type Balance = Balance;
-	type MaxNumberOfTrades = MaxNumberOfTrades;
-	type Currency = MultiInspectAdapter<AccountId, AssetId, Balance, Balances, Tokens, NativeAssetId>;
-	type AMM = (XYK, LBP);
-	type WeightInfo = weights::route_executor::BasiliskWeight<Runtime>;
-}
-
-parameter_types! {
-	pub const RewardPerCollator: Balance = 15_216_000_000_000_000; // 12.68[BSX/block] * 1200[block]
-	//GalacticCouncil collators
-	pub ExcludedCollators: Vec<AccountId> = vec![
-		// bXn5CfJB2qHvqnuMqTpXn6un9Fjch8mwkb9i3JUsGVD4ChLoe
-		hex_literal::hex!["f25e5d7b43266a5b4cca762c9be917f18852d7a5db85e734776206eeb539dd4f"].into(),
-	];
-}
-
-impl pallet_collator_rewards::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type Balance = Balance;
-	type CurrencyId = AssetId;
-	type Currency = Currencies;
-	type RewardPerCollator = RewardPerCollator;
-	type ExcludedCollators = ExcludedCollators;
-	type RewardCurrencyId = NativeAssetId;
-	// We wrap the ` SessionManager` implementation of `CollatorSelection` to get the collatrs that
-	// we hand out rewards to.
-	type SessionManager = CollatorSelection;
-}
-
-// constants need to be in scope to be used in generics
-use pallet_ema_oracle::MAX_PERIODS;
-
-parameter_types! {
-	pub SupportedPeriods: BoundedVec<OraclePeriod, ConstU32<MAX_PERIODS>> = BoundedVec::truncate_from(
-		vec![OraclePeriod::LastBlock, OraclePeriod::Hour, OraclePeriod::Day, OraclePeriod::Week]
-	);
-	// There are currently only a few pools, so the number of entries per block is limited.
-	// NOTE: Needs to be updated once the number of pools grows.
-	pub MaxUniqueOracleEntries: u32 = 30;
-}
-
-impl pallet_ema_oracle::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	type WeightInfo = weights::ema_oracle::BasiliskWeight<Runtime>;
-	type BlockNumberProvider = RelayChainBlockNumberProvider<Runtime>;
-	type SupportedPeriods = SupportedPeriods;
-	type MaxUniqueEntries = MaxUniqueOracleEntries;
 }
 
 // Create the runtime by composing the FRAME pallets that were previously configured.
